Adapted blog post version
This is the adapted version of this enum featured on the blog post "Log Unexpected Switch Options" (http://marxsoftware.blogspot.com/2017/11/log-unexpected-switch-options.html) on the "Inspired by Actual Events" blog (http://marxsoftware.blogspot.com/). The only adaptation from the blog post's version is the package name which was dustin.examples.switchdemo in the blog post and is dustin.utilities.exceptions in this version (the blog post will likely be updated to match this at some point).