Static wrappers catch and handle checked exceptions
Collection of static methods that accept any one of ExceptionThrowingFunction, ExceptionThrowingSupplier, ExceptionThrowingConsumer, and ExceptionThrowingPredicate (each of which can throw a checked exception) and handles any captured exception by re-throwing it as a runtime (unchecked) exception. This allows lambda expressions that throw checked exceptions to be used in JDK 8 stream operations that do not declare throws clauses for checked exceptions.